requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
schedule==1.2.0
requests==2.31.0

# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.0.0